fe: moved fields to subfolder
