+++ frontend-sveltekit/src/lib/components/form/fields/textField.svelte
@@ -1,14 +1,15 @@
 <script lang="ts">
 	import { getContext } from 'svelte';
-	import { formKey } from './form.svelte';
+	import { formKey } from '../form.svelte';
 	import get from 'get-value';
 
-	import FieldWrapper from './fieldParts/fieldWrapper.svelte';
+	import FieldWrapper from '../fieldParts/fieldWrapper.svelte';
 
 	//
 
 	// Field props
 	export let name: string;
+	export let type: 'text' | 'email' | 'password' = 'text';
 	// Label props
 	export let labelText = '';
 	export let labelIcon: Function = null;
@@ -18,6 +19,13 @@
 	// Bottom props
 	export let helperText = '';
 
+	// This is needed to change dynamically the type of the input
+	// Solution provided by
+	// https://github.com/sveltejs/svelte/issues/3921#issuecomment-880664654
+	const setType = (node) => {
+		node.type = type;
+	};
+
 	//
 
 	const { form, errors, handleChange } = getContext(formKey);
@@ -26,13 +34,15 @@
 <!--  -->
 
 <FieldWrapper fieldName={name} {labelText} {labelIcon} {labelLink} {helperText}>
-	<textarea
+	<input
+		use:setType
 		{name}
 		id={name}
+		value={get($form, name)}
 		class="field__input"
 		class:field__input-error={$errors[name]}
 		{placeholder}
 		on:keyup={handleChange}
-		on:blur={handleChange}>{get($form, name)}</textarea
-	>
+		on:blur={handleChange}
+	/>
 </FieldWrapper>
